Add Don Corgi's All-Access Artist Pass offer to README
Hey Dan! 👋

Added my new All-Access Artist Pass for Black Friday.

I'm running a "Small Bet" experiment this year by bundling everything forever for $97 (first time trying a bundle this big!).

If you think this fits the vibe for a shoutout on LinkedIn or on X, I’d really appreciate the boost on this experiment! 🚀

Thanks! 🥰